Moved hub-related views to `FlexHub`
