+++ src/FlexHub/FlexHub.ts
@@ -7,7 +7,7 @@ import Spacer from '@Hi/Components/Spacer';
 import TextView from '@Hi/Components/TextView';
 import VStack from '@Hi/Components/VStack';
 import { ViewController } from '@Hi/ViewController';
-import BrowserPreferences from './BrowserPreferences';
+import BrowserPreferences from '../BrowserPreferences';
 import HubTitlebar from './HubTitlebar';
 
 export default class FlexHub extends HIFullScreenView {
